LayoutTests/plugins/mouse-events-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 0: PLUGIN: mouseDown at (12, 12)
CONSOLE MESSAGE: line 0: PLUGIN: mouseUp at (12, 12)
CONSOLE MESSAGE: line 0: PLUGIN: mouseDown at (22, 22)
CONSOLE MESSAGE: line 0: PLUGIN: mouseUp at (32, 22)

Test for bug 11517: Flash clicks/interactivity not working properly.
